Fix frame indexing

+++ src/python/dicomifier/bruker_to_dicom/image.py
@@ -74,9 +74,10 @@ def _get_pixel_data(data_set, generator, frame_index):
                 data_set["VisuCoreFrameCount"][0]
                 / numpy.cumprod([x[0] for x in non_slice])[-1])
         frame_index = generator.get_linear_index(frame_index)
-        volume = int(frame_index / slices_per_frame)
+        volume = frame_index / slices_per_frame
         slice_index = frame_index % slices_per_frame
-        frame_index = volume*slices_per_frame+(slices_per_frame-slice_index-1)
+        frame_index = int(
+            volume*slices_per_frame+(slices_per_frame-slice_index-1))
     else:
         frame_index = generator.get_linear_index(frame_index)
     frame_data = data_set["PIXELDATA"][frame_index]
